Adding new fonts, icons and new classes #29
I am still in the process of clearing the wrong icon and adding a
suitable ... But these icons are supported by IE 7 and IE 6. will have
to write a js script icons. My question is can you make the default to
run js script in the folder of the skin. For example it said
custom-script.js where everyone can add user script.