Add some visual upgrades! #29
I added the required classes aria-hidden='true'  to display icons as
well. As did a login and improved forum. Still the search box in the
forum to fix it up!
As I fixed and to delete the post, now is very beautiful :)
there is a problem in Internet Explorer buttons in the post
http://f.cl.ly/items/0f1O0P1C1q1S2k1d2X3n/342.jpg
I think we should set its min-width: 85px;
but if anyone has a better solution to share.